wip: lib and bits
