+++ package.json
@@ -34,6 +34,7 @@
 	"dependencies": {
 		"@atem-connection/camera-control": "^0.3.0",
 		"@companion-module/base": "~1.11.3",
+		"@julusian/image-rs": "^1.1.1",
 		"atem-connection": "3.6.0-nightly-master-20250524-142646-fbf40faa.0",
 		"lodash-es": "^4.17.21",
 		"type-fest": "^4.41.0"
